fix: broken dependencies
when merging .tscn files, there can be issues. In this case, the collision shapes didnt import correctly.
The result was that the scene could not be opened at all.
To fix it, i deleted the lines in the .tscn file where the parse error happend and later added them again in godot.